# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Herdwise API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /admins:
    get:
      operationId: listAdmins
      tags: [admins]
      summary: List administrators
      responses:
        "200":
          description: List of administrators
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Admin"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "403":
          description: Access denied
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    post:
      operationId: createAdmin
      tags: [admins]
      summary: Grant administrator access by Clerk user ID
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AdminInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Admin"
        "400":
          description: Invalid input
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "403":
          description: Access denied
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "409":
          description: Already an administrator
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /admins/{id}:
    delete:
      operationId: deleteAdmin
      tags: [admins]
      summary: Revoke administrator access
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Revoked
        "400":
          description: Cannot revoke (e.g. last administrator or self)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "403":
          description: Access denied
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /cattle:
    get:
      operationId: listCattle
      tags: [cattle]
      summary: List all cattle
      parameters:
        - in: query
          name: status
          schema:
            type: string
        - in: query
          name: investorId
          schema:
            type: integer
        - in: query
          name: search
          schema:
            type: string
      responses:
        "200":
          description: List of cattle
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Animal"
    post:
      operationId: createAnimal
      tags: [cattle]
      summary: Add a new animal
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AnimalInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Animal"
  /cattle/{id}:
    get:
      operationId: getAnimal
      tags: [cattle]
      summary: Get animal by ID
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Animal details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Animal"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    patch:
      operationId: updateAnimal
      tags: [cattle]
      summary: Update animal
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AnimalUpdate"
      responses:
        "200":
          description: Updated animal
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Animal"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    delete:
      operationId: deleteAnimal
      tags: [cattle]
      summary: Delete animal
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error
    Animal:
      type: object
      required: [id, tag, breed, status, weightKg, createdAt]
      properties:
        id:
          type: integer
        tag:
          type: string
        previousTag:
          type: ["string", "null"]
        lotNumber:
          type: ["string", "null"]
        breed:
          type: string
        sex:
          type: ["string", "null"]
        stage:
          type: ["string", "null"]
        description:
          type: ["string", "null"]
        status:
          type: string
        weightKg:
          type: ["number", "null"]
        dateReceived:
          type: ["string", "null"]
        notes:
          type: ["string", "null"]
        investorId:
          type: ["integer", "null"]
        investorName:
          type: ["string", "null"]
        createdAt:
          type: string
        marketWeightKg:
          type: number
          description: Target weight (kg) at which the animal is ready for market.
        marketReady:
          type: boolean
          description: True when the current weight is at or above the market target.
        weightRecordCount:
          type: integer
          description: Number of weight records used for the projection.
        avgDailyGainKg:
          type: ["number", "null"]
          description: Average daily weight gain (kg/day) from linear regression over weight history. Null when not computable.
        daysToMarket:
          type: ["integer", "null"]
          description: Projected days until the animal reaches the market weight. 0 if already ready, null when not projectable.
        projectedMarketDate:
          type: ["string", "null"]
          description: Projected calendar date (YYYY-MM-DD) the animal reaches market weight. Null when not projectable.
    AnimalInput:
      type: object
      required: [tag, breed, status]
      properties:
        tag:
          type: string
          minLength: 1
        previousTag:
          type: ["string", "null"]
        lotNumber:
          type: ["string", "null"]
        breed:
          type: string
          minLength: 1
        sex:
          type: ["string", "null"]
        stage:
          type: ["string", "null"]
        description:
          type: ["string", "null"]
        status:
          type: string
        weightKg:
          type: ["number", "null"]
        dateReceived:
          type: ["string", "null"]
        notes:
          type: ["string", "null"]
        investorId:
          type: ["integer", "null"]
    AnimalUpdate:
      type: object
      properties:
        tag:
          type: string
          minLength: 1
        previousTag:
          type: ["string", "null"]
        lotNumber:
          type: ["string", "null"]
        breed:
          type: string
          minLength: 1
        sex:
          type: ["string", "null"]
        stage:
          type: ["string", "null"]
        description:
          type: ["string", "null"]
        status:
          type: string
        weightKg:
          type: ["number", "null"]
        dateReceived:
          type: ["string", "null"]
        notes:
          type: ["string", "null"]
        investorId:
          type: ["integer", "null"]
...
    Admin:
      type: object
      required: [id, clerkUserId, createdAt]
      properties:
        id:
          type: integer
        clerkUserId:
          type: string
        email:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time
        isSelf:
          type: boolean
    AdminInput:
      type: object
      required: [clerkUserId]
      properties:
        clerkUserId:
          type: string
          minLength: 1
        email:
          type: ["string", "null"]
...