# lib/api-zod/src/generated/types/admin.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Herdwise API specification
 * OpenAPI spec version: 0.1.0
 */

export interface Admin {
  id: number;
  clerkUserId: string;
  /** @nullable */
  email?: string | null;
  createdAt: Date;
  isSelf?: boolean;
}